Add fab to note list screen
