+++ app/src/main/java/com/jshvarts/todoapp/MainActivity.kt
@@ -1,52 +1,22 @@
 package com.jshvarts.todoapp
 
-import android.annotation.SuppressLint
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.annotation.StringRes
-import androidx.compose.animation.core.animateFloatAsState
-import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material.*
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Delete
-import androidx.compose.material.pullrefresh.PullRefreshIndicator
-import androidx.compose.material.pullrefresh.pullRefresh
-import androidx.compose.material.pullrefresh.rememberPullRefreshState
-import androidx.compose.runtime.*
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.wrapContentSize
+import androidx.compose.material.CircularProgressIndicator
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.rotate
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.compose.rememberNavController
-import com.google.accompanist.pager.ExperimentalPagerApi
-import com.google.accompanist.pager.HorizontalPager
-import com.google.accompanist.pager.pagerTabIndicatorOffset
-import com.google.accompanist.pager.rememberPagerState
-import com.jshvarts.todoapp.data.Note
-import com.jshvarts.todoapp.notedetail.NoteDetailViewModel
-import com.jshvarts.todoapp.notedetail.ui.NoteDetailUiAction
-import com.jshvarts.todoapp.notedetail.ui.NoteDetailUiState
-import com.jshvarts.todoapp.notelist.NoteListViewModel
-import com.jshvarts.todoapp.notelist.ui.NoteListTodosUiState
-import com.jshvarts.todoapp.notelist.ui.NoteListUiAction
-import com.jshvarts.todoapp.notelist.ui.NoteListUiEffect
-import com.jshvarts.todoapp.notelist.ui.NoteListUiState
 import com.jshvarts.todoapp.ui.navigation.NotesAppNavHost
 import com.jshvarts.todoapp.ui.theme.ToDoAppTheme
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
@@ -61,62 +31,6 @@ class MainActivity : ComponentActivity() {
         }
       }
     }
-  }
-}
-
-@SuppressLint("UnusedMaterialScaffoldPaddingParameter")
-@Composable
-fun NoteListScreen(
-  onNoteClick: (Int) -> Unit,
-  modifier: Modifier = Modifier,
-  viewModel: NoteListViewModel = hiltViewModel()
-) {
-  val uiState by viewModel.uiState.collectAsStateWithLifecycle()
-  val uiEffect by viewModel.uiEffect.collectAsStateWithLifecycle(NoteListUiEffect.Initial)
-  val scaffoldState = rememberScaffoldState()
-  val refreshFailedMessage = stringResource(id = R.string.refresh_failed_message)
-  val deleteFailedMessage = stringResource(id = R.string.delete_failed_message)
-  val scope = rememberCoroutineScope()
-
-  LaunchedEffect(scaffoldState.snackbarHostState) {
-    scope.launch {
-      when (uiEffect) {
-        NoteListUiEffect.RefreshFailed -> {
-          viewModel.uiEffect.collect {
-            scaffoldState.snackbarHostState.showSnackbar(message = refreshFailedMessage)
-          }
-        }
-        NoteListUiEffect.DeleteFailed -> {
-          viewModel.uiEffect.collect {
-            scaffoldState.snackbarHostState.showSnackbar(message = deleteFailedMessage)
-          }
-        }
-      }
-    }
-  }
-
-  Scaffold(
-    scaffoldState = scaffoldState,
-    topBar = {
-      TopAppBar(
-        title = {
-          Text(text = stringResource(id = R.string.notes))
-        },
-        backgroundColor = Color.White,
-        contentColor = Color.Black
-      )
-    },
-    modifier = modifier
-  ) {
-    NoteListSuccessState(
-      uiState = uiState,
-      onNoteClick = onNoteClick,
-      onSwipeToDelete = { note ->
-        viewModel.dispatchAction(
-          NoteListUiAction.SwipeToDelete(note.id)
-        )
-      }
-    )
   }
 }
 
@@ -144,245 +58,4 @@ fun ErrorState(
   }
 }
 
-@OptIn(ExperimentalPagerApi::class, ExperimentalMaterialApi::class)
-@Composable
-fun NoteListSuccessState(
-  uiState: NoteListUiState,
-  onNoteClick: (Int) -> Unit,
-  onSwipeToDelete: (Note) -> Unit,
-  viewModel: NoteListViewModel = hiltViewModel(),
-  modifier: Modifier = Modifier
-) {
-  var tabIndex by remember { mutableStateOf(0) }
-  val tabTitles = listOf(
-    R.string.pending_todos,
-    R.string.completed_todos
-  )
-  val pagerState = rememberPagerState()
-  val scope = rememberCoroutineScope()
-
-  val refreshScope = rememberCoroutineScope()
-  var refreshing by remember { mutableStateOf(false) }
-
-  fun refresh() = refreshScope.launch {
-    refreshing = true
-    delay(300)
-    viewModel.dispatchAction(NoteListUiAction.PullToRefresh)
-    refreshing = false
-  }
-
-  val pullRefreshState = rememberPullRefreshState(refreshing, ::refresh)
-
-  Column(modifier = modifier) {
-    TabRow(
-      backgroundColor = Color.White,
-      contentColor = Color.Black,
-      selectedTabIndex = tabIndex,
-      indicator = { tabPositions ->
-        TabRowDefaults.Indicator(
-          Modifier.pagerTabIndicatorOffset(
-            pagerState,
-            tabPositions,
-          )
-        )
-      }) {
-      tabTitles.forEachIndexed { index, titleResId ->
-        Tab(selected = tabIndex == index,
-          onClick = {
-            tabIndex = index
-            scope.launch {
-              pagerState.animateScrollToPage(index)
-            }
-          },
-
-          text = { Text(stringResource(id = titleResId)) })
-      }
-    }
-
-    val emptyTodosResId = if (pagerState.currentPage == 0) R.string.empty_pending_todos else R.string.empty_completed_todos
-
-    HorizontalPager(
-      count = tabTitles.size,
-      state = pagerState,
-      userScrollEnabled = false
-    ) { tabIndex ->
-      Box(Modifier.pullRefresh(pullRefreshState)) {
-        if (tabIndex == 0) {
-          when (uiState.pendingTodosUiState) {
-            is NoteListTodosUiState.Error -> ErrorState(uiState.pendingTodosUiState.throwable?.message.orEmpty())
-            NoteListTodosUiState.Loading -> LoadingState()
-            is NoteListTodosUiState.Success -> {
-              TodosSuccessState(
-                data = uiState.pendingTodosUiState.data,
-                emptyTodosResId = emptyTodosResId,
-                onNoteClick = onNoteClick,
-                onSwipeToDelete = onSwipeToDelete
-              )
-            }
-          }
-        } else {
-          when (uiState.completedTodosUiState) {
-            is NoteListTodosUiState.Error -> ErrorState(uiState.completedTodosUiState.throwable?.message.orEmpty())
-            NoteListTodosUiState.Loading -> LoadingState()
-            is NoteListTodosUiState.Success -> {
-              TodosSuccessState(
-                data = uiState.completedTodosUiState.data,
-                emptyTodosResId = emptyTodosResId,
-                onNoteClick = onNoteClick,
-                onSwipeToDelete = onSwipeToDelete
-              )
-            }
-          }
-        }
-        PullRefreshIndicator(refreshing, pullRefreshState, Modifier.align(Alignment.TopCenter))
-      }
-    }
-  }
-}
-
-@OptIn(ExperimentalMaterialApi::class)
-@Composable
-fun TodosSuccessState(
-  data: List<Note>,
-  @StringRes emptyTodosResId: Int,
-  onNoteClick: (Int) -> Unit,
-  onSwipeToDelete: (Note) -> Unit,
-  modifier: Modifier = Modifier,
-) {
-  if (data.isEmpty()) {
-    TodosEmptyState(resId = emptyTodosResId)
-  } else {
-    LazyColumn {
-      items(
-        items = data,
-        key = { it.id }
-      ) { item ->
-        val dismissState = rememberDismissState()
-        val isDismissed = dismissState.isDismissed(DismissDirection.EndToStart)
-        if (isDismissed) {
-          onSwipeToDelete(item)
-        }
-
-        val degrees by animateFloatAsState(
-          if (dismissState.targetValue == DismissValue.Default) 0f else -45f
-        )
-
-        SwipeToDismiss(
-          state = dismissState,
-          directions = setOf(DismissDirection.EndToStart),
-          dismissThresholds = { FractionalThreshold(0.2f) },
-          background = { SwipeToDismissBackground(degrees = degrees) },
-          dismissContent = {
-            NoteListItem(
-              item = item,
-              onNoteClick = onNoteClick
-            )
-          },
-          modifier = modifier
-        )
-      }
-    }
-  }
-}
-
-@Composable
-fun SwipeToDismissBackground(degrees: Float) {
-  Box(
-    modifier = Modifier
-      .fillMaxSize()
-      .background(Color.White)
-      .padding(horizontal = 16.dp),
-    contentAlignment = Alignment.CenterEnd
-  ) {
-    Icon(
-      modifier = Modifier
-        .rotate(degrees),
-      imageVector = Icons.Filled.Delete,
-      contentDescription = stringResource(id = R.string.delete_icon),
-      tint = Color.White
-    )
-  }
-}
-
-@Composable
-fun NoteListItem(
-  item: Note,
-  onNoteClick: (Int) -> Unit,
-  modifier: Modifier = Modifier
-) {
-  Row(
-    horizontalArrangement = Arrangement.spacedBy(16.dp),
-    modifier = modifier
-      .fillMaxWidth()
-      .clickable {
-        onNoteClick.invoke(item.id)
-      }
-  ) {
-    Text(
-      text = item.title,
-      style = MaterialTheme.typography.h6,
-      modifier = modifier
-        .padding(16.dp)
-    )
-  }
-}
-
-@Composable
-fun TodosEmptyState(@StringRes resId: Int) {
-  Box(
-    modifier = Modifier
-      .fillMaxSize()
-      .wrapContentSize(Alignment.Center)
-  ) {
-    Text(
-      text = stringResource(id = resId),
-      style = MaterialTheme.typography.h6,
-      textAlign = TextAlign.Center,
-      modifier = Modifier.fillMaxWidth()
-    )
-  }
-}
-
-@Composable
-fun NoteDetailScreen(
-  noteId: Int,
-  modifier: Modifier = Modifier,
-  viewModel: NoteDetailViewModel = hiltViewModel()
-) {
-  val uiState by viewModel.uiState.collectAsStateWithLifecycle()
-
-  LaunchedEffect(Unit) {
-    viewModel.dispatchAction(NoteDetailUiAction.LoadNote(noteId))
-  }
-
-  when (uiState) {
-    NoteDetailUiState.Loading -> LoadingState()
-    is NoteDetailUiState.Success -> NoteDetailSuccessState(
-      uiState as NoteDetailUiState.Success,
-    )
-    is NoteDetailUiState.Error -> ErrorState((uiState as NoteDetailUiState.Error).throwable?.message.orEmpty())
-  }
-}
-
-@Composable
-fun NoteDetailSuccessState(
-  uiState: NoteDetailUiState.Success,
-  modifier: Modifier = Modifier
-) {
-  Column(
-    verticalArrangement = Arrangement.spacedBy(16.dp),
-    modifier = modifier
-      .fillMaxSize()
-      .padding(16.dp)
-  ) {
-    Text(
-      text = "ID: ${uiState.note.id}",
-      style = MaterialTheme.typography.h3,
-    )
-    Text(
-      text = "Text: ${uiState.note.title}",
-      style = MaterialTheme.typography.h3,
-    )
-  }
-}
