+++ app/src/main/java/com/jshvarts/todoapp/data/NoteRepository.kt
@@ -7,4 +7,5 @@ interface NoteRepository {
   fun getNote(id: Int): Flow<Note>
   suspend fun refreshNotes(): Result<Unit>
   suspend fun deleteNote(id: Int): Result<Unit>
+  suspend fun addNote(note: Note): Result<Unit>
 }
