+++ app/src/main/java/com/jshvarts/todoapp/data/local/NoteIdFactory.kt
@@ -1,0 +1,5 @@
+package com.jshvarts.todoapp.data.local
+
+interface NoteIdFactory {
+  suspend fun getNextAvailableId(): Int
+}
