+++ app/src/main/java/com/jshvarts/todoapp/data/local/NoteIdFactoryImpl.kt
@@ -1,0 +1,14 @@
+package com.jshvarts.todoapp.data.local
+
+import javax.inject.Inject
+import javax.inject.Singleton
+
+@Singleton
+class NoteIdFactoryImpl @Inject constructor(
+  val notesDao: NotesDao
+) : NoteIdFactory {
+  override suspend fun getNextAvailableId(): Int {
+    val currentMaxNoteId = notesDao.getCurrentMaxNoteId()
+    return currentMaxNoteId + 1
+  }
+}
