+++ app/src/main/java/com/jshvarts/todoapp/di/AppModule.kt
@@ -2,6 +2,8 @@ package com.jshvarts.todoapp.di
 
 import com.jshvarts.todoapp.data.NoteRepository
 import com.jshvarts.todoapp.data.OfflineFirstNoteRepository
+import com.jshvarts.todoapp.data.local.NoteIdFactory
+import com.jshvarts.todoapp.data.local.NoteIdFactoryImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -12,4 +14,7 @@ import dagger.hilt.components.SingletonComponent
 abstract class AppModule {
   @Binds
   abstract fun bindsNotesRepository(offlineFirstNoteRepository: OfflineFirstNoteRepository): NoteRepository
+
+  @Binds
+  abstract fun bindsNoteIdFactory(noteIdFactory: NoteIdFactoryImpl): NoteIdFactory
 }
