+++ app/src/main/java/com/jshvarts/todoapp/notedetail/ui/NoteUiContract.kt
@@ -9,6 +9,7 @@ import kotlinx.parcelize.Parcelize
 
 sealed interface NoteDetailUiAction : UiAction {
   data class LoadNote(val id: Int) : NoteDetailUiAction
+  object NewNote : NoteDetailUiAction
 }
 
 @Parcelize
