+++ app/src/main/res/values/strings.xml
@@ -8,4 +8,5 @@
     <string name="refresh_failed_message">Refresh failed. Please try again later.</string>
     <string name="delete_icon">Delete todo</string>
     <string name="delete_failed_message">Delete failed. Please try again.</string>
+    <string name="add_note">Add note</string>
 </resources>
